+++ src/app/app-routing.module.ts
@@ -3,7 +3,7 @@ import { RouterModule, Routes } from '@angular/router';
 import { LoginComponent } from './login/login.component';
 import { RegisterComponent } from './register/register.component';
 import { HomeComponent } from './home/home.component';
-import { HOME, LOGIN, REGISTER } from './Models/contants';
+import { HOME, LOGIN, REGISTER } from './Models/constants';
 
 const routes: Routes = [
   { path: '', component: HomeComponent },
